# [FLINK-16404][runtime] Avoid caching buffers for blocked input channels before barrier alignment