--- a/flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/output/operators/BroadcastStateBootstrapOperator.java
+++ b/flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/output/operators/BroadcastStateBootstrapOperator.java
@@ -74,6 +74,8 @@ public class BroadcastStateBootstrapOperator<IN>
 			this,
 			getRuntimeContext().getIndexOfThisSubtask(),
 			timestamp,
+			getContainingTask().getConfiguration().isExactlyOnceCheckpointMode(),
+			getContainingTask().getConfiguration().isUnalignedCheckpointsEnabled(),
 			getContainingTask().getCheckpointStorage(),
 			savepointPath);