--- a/flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/output/operators/KeyedStateBootstrapOperator.java
+++ b/flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/output/operators/KeyedStateBootstrapOperator.java
@@ -86,6 +86,8 @@ public class KeyedStateBootstrapOperator<K, IN>
 			this,
 			getRuntimeContext().getIndexOfThisSubtask(),
 			timestamp,
+			getContainingTask().getConfiguration().isExactlyOnceCheckpointMode(),
+			getContainingTask().getConfiguration().isUnalignedCheckpointsEnabled(),
 			getContainingTask().getCheckpointStorage(),
 			savepointPath);