--- a/flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/output/operators/StateBootstrapOperator.java
+++ b/flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/output/operators/StateBootstrapOperator.java
@@ -71,6 +71,8 @@ public class StateBootstrapOperator<IN>
 			this,
 			getRuntimeContext().getIndexOfThisSubtask(),
 			timestamp,
+			getContainingTask().getConfiguration().isExactlyOnceCheckpointMode(),
+			getContainingTask().getConfiguration().isUnalignedCheckpointsEnabled(),
 			getContainingTask().getCheckpointStorage(),
 			savepointPath);