--- a/flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/FailedCheckpointStats.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/FailedCheckpointStats.java
@@ -41,9 +41,6 @@ public class FailedCheckpointStats extends AbstractCheckpointStats {
 	/** Total checkpoint state size over all subtasks. */
 	private final long stateSize;
 
-	/** Buffered bytes during alignment over all subtasks. */
-	private final long alignmentBuffered;
-
 	/** Timestamp when the checkpoint was failed at the coordinator. */
 	private final long failureTimestamp;
 
@@ -68,7 +65,6 @@ public class FailedCheckpointStats extends AbstractCheckpointStats {
 	 * @param taskStats Task stats for each involved operator.
 	 * @param numAcknowledgedSubtasks Number of acknowledged subtasks.
 	 * @param stateSize Total checkpoint state size over all subtasks.
-	 * @param alignmentBuffered Buffered bytes during alignment over all subtasks.
 	 * @param failureTimestamp Timestamp when this checkpoint failed.
 	 * @param latestAcknowledgedSubtask The latest acknowledged subtask stats or <code>null</code>.
 	 * @param cause Cause of the checkpoint failure or <code>null</code>.
@@ -81,7 +77,6 @@ public class FailedCheckpointStats extends AbstractCheckpointStats {
 			Map<JobVertexID, TaskStateStats> taskStats,
 			int numAcknowledgedSubtasks,
 			long stateSize,
-			long alignmentBuffered,
 			long failureTimestamp,
 			@Nullable SubtaskStateStats latestAcknowledgedSubtask,
 			@Nullable Throwable cause) {
@@ -91,7 +86,6 @@ public class FailedCheckpointStats extends AbstractCheckpointStats {
 		this.numAcknowledgedSubtasks = numAcknowledgedSubtasks;
 		checkArgument(stateSize >= 0, "Negative state size");
 		this.stateSize = stateSize;
-		this.alignmentBuffered = alignmentBuffered;
 		this.failureTimestamp = failureTimestamp;
 		this.latestAcknowledgedSubtask = latestAcknowledgedSubtask;
 		this.failureMsg = cause != null ? cause.getMessage() : null;
@@ -110,11 +104,6 @@ public class FailedCheckpointStats extends AbstractCheckpointStats {
 	@Override
 	public long getStateSize() {
 		return stateSize;
-	}
-
-	@Override
-	public long getAlignmentBuffered() {
-		return alignmentBuffered;
 	}
 
 	@Override