--- a/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/NetworkClientHandler.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/NetworkClientHandler.java
@@ -49,4 +49,11 @@ public interface NetworkClientHandler extends ChannelHandler {
 	 * @param inputChannel The input channel with unannounced credits.
 	 */
 	void notifyCreditAvailable(final RemoteInputChannel inputChannel);
+
+	/**
+	 * Resumes data consumption from the producer after an exactly once checkpoint.
+	 *
+	 * @param inputChannel The input channel to resume data consumption.
+	 */
+	void resumeConsumption(RemoteInputChannel inputChannel);
 }